Refactored Binder Reuse tests to use Antlr based parser.
