+++ IronVelocity.Tests/TemplateExecution/BinderReuse/SetMemberReuseTests.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using IronVelocity.Tests;
 
-namespace IronVelocity.Tests.Binders.Reuse
+namespace IronVelocity.Tests.TemplateExecution.BinderReuse
 {
     [TestFixture]
     public class SetMemberReuseTests : BinderReuseTestBase
@@ -19,17 +19,17 @@ namespace IronVelocity.Tests.Binders.Reuse
             for (int i = 0; i < 5; i++)
             {
                 var helper = new SetMemberHelper();
-                var context = new Dictionary<string, object>{
-                    {"x", helper}
-                };
+                var context = new { x = helper };
 
                 var input = "#set($x.Property = 3)\r\n#set($x.Property = 123)Hello World";
 
-                Utility.TestExpectedMarkupGenerated(input, expectedOutput, context, isGlobalEnvironment: false);
-                Assert.AreEqual(123, helper.Property);
+                var execution = ExecuteTemplate(input, context);
+
+                Assert.That(execution.Output, Is.EqualTo(expectedOutput));
+                Assert.That(helper.Property, Is.EqualTo(123));
 
             }
-            Assert.AreEqual(1, CallSiteBindCount);
+            Assert.That(CallSiteBindCount, Is.EqualTo(1));
         }
 
         public class SetMemberHelper
